Update test_add_block to work around an issue with scripts not executing.
I tracked down this issue to being caused by scripts not publishing their initial state.
It is not clear to me why this is happening primarily in this unit test and not affecting other unit tests which are very similar to it.
I at least managed to determine it is not related to the block feature.
Also, even though this issue is more frequent in our Jenkins server and on my linux dev machine, I have not been able to reproduce it in my laptop (mac).
Would be interesting to see how this is going to behave in the wild, e.g. TTS, BTS and summit.